requirements.txt
mcp>=1.0.0
PyGithub>=2.1.1
py-trello>=0.19.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.0.0
anthropic>=0.7.0
tiktoken>=0.5.0 